Move Secret to utils
