+++ pkg/utils/utils.go
@@ -40,3 +40,27 @@ func HashSHA(s string) []byte {
 func Path(keys ...string) string {
 	return strings.Join(append([]string{}, keys...), "/")
 }
+
+// Secret special type for storing secrets.
+type Secret string
+
+// MarshalYAML implements the yaml.Marshaler interface for Secrets.
+func (s Secret) MarshalYAML() (interface{}, error) {
+	if s != "" {
+		return "<secret>", nil
+	}
+	return nil, nil
+}
+
+// UnmarshalYAML implements the yaml.Unmarshaler interface for Secrets.
+func (s *Secret) UnmarshalYAML(unmarshal func(interface{}) error) error {
+	type plain Secret
+	return unmarshal((*plain)(s))
+}
+
+func (s Secret) MarshalJSON() ([]byte, error) {
+	if s != "" {
+		return []byte(`"<secret>"`), nil
+	}
+	return nil, nil
+}
